MNOGO TESTOB

+++ src/cw/main_cycle/parse_args.c
@@ -6,14 +6,14 @@
 /*   By: bharrold <bharrold@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/03 21:24:14 by bharrold          #+#    #+#             */
-/*   Updated: 2020/02/03 22:23:39 by bharrold         ###   ########.fr       */
+/*   Updated: 2020/02/12 05:20:44 by bharrold         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "corewar.h"
 #include "cw_ops.h"
 
-static void	set_arg_type(int8_t arg_type, int index, t_carry *carry)
+static void	set_arg_type(int8_t arg_type, int8_t index, t_carry *carry)
 {
 	carry->args_types[index - 1] = g_arg_code[arg_type - 1];
 }
@@ -33,11 +33,14 @@ void		clear_args(t_cw *cw, t_carry *carry)
 
 void		parse_args(t_cw *cw, t_carry *carry, t_op *op)
 {
-	uint8_t args_types_code;
+	int8_t args_types_code;
 	
 	if (op->args_types_code)
 	{
 		args_types_code = arena_get_byte(cw, carry->position, 1);
+		// if (cw->cycles == 2390) {
+		// 	printf("byte %x\n", args_types_code);
+		// }
 		if (op->args_num >= 1)
 			set_arg_type((int8_t)((args_types_code & 0xC0) >> 6), 1, carry);
 		if (op->args_num >= 2)
@@ -47,4 +50,7 @@ void		parse_args(t_cw *cw, t_carry *carry, t_op *op)
 	}
 	else
 		carry->args_types[0] = op->args_types[0];
+	// if (cw->cycles == 2390) {
+	// 	printf("args2 %d %d %d\n", carry->args_types[0], carry->args_types[1], carry->args_types[2]);
+	// }
 }
